Create an utils methods service
